+++ service/src/main/java/uk/gov/mca/beacons/api/legacybeacon/application/LegacyBeaconService.java
@@ -124,15 +124,15 @@ public class LegacyBeaconService {
     return legacyBeaconRepository.saveAll(legacyBeacons);
   }
 
-  public Map<String, Long> findHexIdsWithDuplicates() {
-    Map<String, Long> hexIdsAndDuplicateCounts = legacyBeaconRepository
+  public Map<String, Integer> findHexIdsWithDuplicates() {
+    Map<String, Integer> hexIdsAndDuplicateCounts = legacyBeaconRepository
       .findByHexIdNotNull()
       .stream()
       .collect(groupingBy(LegacyBeacon::getHexId, Collectors.counting()))
       .entrySet()
       .stream()
       .filter(m -> m.getValue() > 1)
-      .collect(Collectors.toMap(m -> m.getKey(), m -> m.getValue()));
+      .collect(Collectors.toMap(m -> m.getKey(), m -> m.getValue().intValue()));
     return hexIdsAndDuplicateCounts;
   }
